# specs/001-streamwise/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StreamWise API
  version: 0.1.0
  description: |
    REST API for StreamWise — movie/series discovery, interactions, and personalized recommendations.
    Brazil (BR) streaming availability. See specs/001-streamwise/plan.md for architecture.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /catalog/trending:
    get:
      tags: [catalog]
      summary: Trending movies and series
      operationId: getTrending
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [movie, series, all]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: Trending titles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleListResponse"
  /catalog/new:
    get:
      tags: [catalog]
      summary: New releases and on-air series
      operationId: getNewReleases
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: New titles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleListResponse"
...
  /catalog/search:
    get:
      tags: [catalog]
      summary: Natural language / keyword title search
      operationId: searchTitles
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: provider_ids
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
          style: form
          explode: true
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleListResponse"
...
  /catalog/genres:
    get:
      tags: [catalog]
      summary: List genres for onboarding and filters
      operationId: listGenres
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Genre options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreListResponse"
  /catalog/providers:
    get:
      tags: [catalog]
      summary: List streaming providers (Brazil)
      operationId: listProviders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Provider options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderListResponse"
  /catalog/changes:
    get:
      tags: [catalog]
      summary: Recent catalog availability changes
      operationId: getCatalogChanges
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: Catalog diff events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogChangeListResponse"
...
  /integrations/trakt/authorize:
    get:
      tags: [integrations]
      summary: Trakt OAuth authorize URL (stub)
      operationId: traktAuthorize
      responses:
        "200":
          description: Authorize URL
        "503":
          description: Trakt not configured
  /integrations/trakt/callback:
    get:
      tags: [integrations]
      summary: Trakt OAuth callback (stub)
      operationId: traktCallback
      parameters:
        - name: code
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Stub response
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
...
  schemas:
...
    TitleSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tmdb_id:
          type: integer
        type:
          type: string
          enum: [movie, series]
        title:
          type: string
        overview:
          type: string
        release_date:
          type: string
          format: date
        poster_url:
          type: string
        streamwise_avg_rating:
          type: number
          nullable: true
        like_count:
          type: integer
        genres:
          type: array
          items:
            type: string
        streaming_providers:
          type: array
          items:
            $ref: "#/components/schemas/StreamingProviderBadge"
...
    StreamingProviderBadge:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logo_url:
          type: string
        availability_type:
          type: string
          enum: [flatrate, rent, buy]
    TitleListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TitleSummary"
        total:
          type: integer
...
    GenreListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
    ProviderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              logo_url:
                type: string
                nullable: true
    CatalogChangeListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CatalogChangeItem"
        total:
          type: integer
    CatalogChangeItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title_id:
          type: string
          format: uuid
        title_name:
          type: string
        provider_name:
          type: string
        change_type:
          type: string
        availability_type:
          type: string
        detected_at:
          type: string
          format: date-time
...